src/requirements.txt
Python 3.8.12
conda==4.10.3
numpy==1.19.5
pandas==1.3.4
matplotlib==3.4.3
seaborn==0.11.2
Pillow==8.4.0
opencv-python==4.5.4.58
scikit-image==0.18.3
torch==1.10.0
torchvision==0.11.1
pytorch-gradcam==0.2.1
tensorflow==2.6.0
keras==2.6.0
mlflow==1.20.2
Flask==2.0.2
streamlit==1.1.0